+++ src/models/snipcart/Order.php
@@ -16,7 +16,7 @@ use workingconcept\snipcart\helpers\ModelHelper;
  * Snipcart Order model
  * https://docs.snipcart.com/v2/api-reference/orders
  *
- * @package workingconcept\snipcart\models
+ * @package workingconcept\snipcart\models\snipcart
  *
  * @property Address $billingAddress
  * @property Address $shippingAddress
@@ -410,6 +410,11 @@ class Order extends \craft\base\Model
     public $totalPriceWithoutDiscountsAndTaxes;
 
     /**
+     * @var
+     */
+    public $paymentDetails;
+
+    /**
      * @return array
      */
     public function getDiscounts(): array
@@ -692,7 +697,7 @@ class Order extends \craft\base\Model
             'shippingAddressProvince',
             'shippingAddressPostalCode',
             'shippingAddressPhone',
-            'cpUrl',
+            'cpUrl'
         ];
     }